fix(runtime): fix compile error or warning in web assembly

+++ src/runtime/vm/rt_string.cpp
@@ -175,7 +175,7 @@ RtString* String::fast_allocate_string(int32_t length)
     // String::GetLegacyNonRandomizedHashCode need zero terminated string, so we allocate one extra character.
     // TODO: can we optimize it out? we have redirected String::GetHashCode and String::GetLegacyNonRandomizedHashCode to
     // the intrinsic implementation which does not require zero-termination.
-    size_t allocation_size = get_string_allocation_size(length);
+    size_t allocation_size = static_cast<size_t>(get_string_allocation_size(length));
     #if LEANCLR_GC_DEBUG
     RtString* newString = (RtString*)gc::GarbageCollector::allocate_object(g_stringClass, allocation_size,
                                                                   ::leanclr::gc::GcAllocSite::make_internal(__FILE__, __LINE__, "String::fast_allocate_string"));
